use user supplied threshold in MatchTemplate

+++ label/MatchTemplate.h
@@ -49,8 +49,8 @@ public:
 	MatchTemplate();
 	virtual ~MatchTemplate();
 
-	static vector<tMatch> *matchTemplate(Mat templ, double area, map<int, tFrame*> frames, int method, int startingFrame, int numFrames, int step, int threshold = 200000);
-	static vector<tMatch> *matchTemplate(vector<tMatch> *matches, map<int, tFrame*> frames, int method, int startingFrame, int numFrames, int step, int threshold = 30000);
+	static vector<tMatch> *matchTemplate(Mat templ, double area, map<int, tFrame*> frames, int method, int startingFrame, int numFrames, int step, double threshold = 200000);
+	static vector<tMatch> *matchTemplate(vector<tMatch> *matches, map<int, tFrame*> frames, int method, int startingFrame, int numFrames, int step, double threshold = 30000);
 
 	static void displayMatches(vector<tMatch> *matches);
